[#AC-14](fix) 아에팀 추가 vm 구현 (데이터형 double > BigDecimal로 수정)
- 원인 : 소수점 계산시 double은 부정확한 데이터가 나올 수 있음.
- 수정 내용 : 정확한 소수점 계산을 위한 bigDecimal로 데이터형 수정 https://dev.gmarket.com/75

+++ core/db/src/main/kotlin/com/hj/average/core/db/entity/ItemEntity.kt
@@ -2,6 +2,7 @@ package com.hj.average.core.db.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigDecimal
 
 @Entity
 data class ItemEntity(
@@ -9,8 +10,8 @@ data class ItemEntity(
     val id: Int = 0,
     val name: String,
     val date: Long,
-    val firstPrice: Double,
-    val firstQuantity: Double,
-    val secondPrice: Double,
-    val secondQuantity: Double
+    val firstPrice: BigDecimal,
+    val firstQuantity: BigDecimal,
+    val secondPrice: BigDecimal,
+    val secondQuantity: BigDecimal
 )
